Update to Rules and Regulations (2.1.8) and (2.7.4) (#1)
Changes to the Rules and Regulations as follows: 
(2.1.8) If an officer is involved with an organization that is actively renting items from Lens and Lights, said officer must lock their officer key ring in the lock box for the duration of the rental period.

(2.7.4) The Vice President or their designee must give approval for additional items to be rented following the original rental load out. Under no condition may items be taken without prior approval being granted. Any item taken without permission may be recalled at the discretion of the Executive Board.